Initial commit
Commits the blender and .obj file of the right arm which has also been
UV unwrapped. ALso commits example files for using the potpurri3d
python library to find geodesic distances on a UV sphere and the right
arm mesh. There still needs to be code to use the UV mapping to select
the geodesic path verticies.