fix: Corrects right and left arm images/uv maps
The original images were downsampled and gave borders to the images.
This created two sepereate issues. the first was that the pixel measures
from location mapping to the uv maps would be vastly off. The second was
that the aspect ratio of the UV maps was wrong as the borders to the
images was unequal in the x and y directions.