feat: Adds calculation of surface area of location
Takes a drawn border of a location in 2D and finds the surface area of
that same location in 3D. NOTE: The units in 3D have yet to be
normalized to something so they are kind of meaningless at this point.
My suggestion is to normalize the 2D drawings by the total possible
drawing area on the subject drawings and normalize the 3D areas by the
total surface area of the mesh as there is a 1:1 mapping from the
drawings to the mesh.